check input and export files for correctness

+++ Calculator/history.py
@@ -1,5 +1,6 @@
 import json, csv
 from Calculator.history_entry import HistoryEntry
+from Calculator.files import is_filename
 
 class historyIterator:
     def __init__(self, items):
@@ -41,9 +42,7 @@ class history:
         for c in self:
             print(c.symbol+" "+str(c.op_value)+" ") 
 
-    def remove_id_from_history(self):
-        self.print_history_commands()
-        history_id_remove = input("Which ID of the history should be removed? ")
+    def remove_id_from_history(self, history_id_remove):
         for idx, calculation in enumerate(self):
             if calculation.id == int(history_id_remove):
                 del self.__calc_history[idx]
@@ -51,9 +50,10 @@ class history:
         return self
     
     def save_history(self, json_file_name):
-        with open(json_file_name, "w", encoding="utf-8") as history_file:
-            entries_as_dicts = [entry.to_dict() for entry in self]
-            json.dump(entries_as_dicts, history_file, indent=2)
+        if is_filename(json_file_name):
+            with open(json_file_name, "w", encoding="utf-8") as history_file:
+                entries_as_dicts = [entry.to_dict() for entry in self]
+                json.dump(entries_as_dicts, history_file, indent=2)
 
     def load_history(self, json_file_name):
         try:
@@ -61,7 +61,7 @@ class history:
                 loaded_data = json.load(history_file)
                 # Convert dicts back to HistoryEntry objects
                 self.__calc_history = [
-                    HistoryEntry(entry["id"], entry["op_name"], entry["symbol"], entry["op_value"])
+                    HistoryEntry(entry["id"], entry["command"], entry["symbol"], entry["operand"])
                     for entry in loaded_data
                 ]
                 # Update history_id to the highest existing ID
@@ -73,11 +73,12 @@ class history:
             print(f"Error reading history file {json_file_name}. Starting with empty history.")
     
     def save_history_csv(self, csv_file_name):
-        with open(csv_file_name, "w", newline="") as f:
-            writer = csv.DictWriter(f, fieldnames=["id", "command", "symbol", "operand"])
-            writer.writeheader()
-            entries_as_dicts = [entry.to_dict() for entry in self]
-            writer.writerows(entries_as_dicts)
+        if is_filename(csv_file_name):
+            with open(csv_file_name, "w", newline="") as f:
+                writer = csv.DictWriter(f, fieldnames=["id", "command", "symbol", "operand"])
+                writer.writeheader()
+                entries_as_dicts = [entry.to_dict() for entry in self]
+                writer.writerows(entries_as_dicts)
